[plymouth-theme-droidian] message: handle whitespaced lines where the last space is in the first half
The split_in_two_lines() helper function splits a string in two lines. It works by looking at the
latter half of the string, and the split happens at the first whitespace there.

This means that

    Installing feature bundle droidian adaptation fxtec pro1

gets correctly split in

    Installing feature bundle droidian
    adaptation fxtec pro1

while

    Installing feature bundle droidian-adaptation-fxtec-pro1

does not (since there aren't any spaces in the latter half).

Handle this case by keeping track of every passed space and trigger
the line split when we have reached the last space.

The example above would be split to

    Installing feature bundle
    droidian-adaptation-fxtec-pro1

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>